fix: Harmonize GUI implementation
